:recycle: Refactor board functions
